# src/Trips.Api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Trips.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5099/"
    }
  ],
  "paths": {
...
    "/trips/{id}": {
      "get": {
        "tags": [
          "Trips"
        ],
        "operationId": "GetTrip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Trips"
        ],
        "operationId": "DeleteTrip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/trips/{tripId}/participants/{pid}": {
      "delete": {
        "tags": [
          "Participants"
        ],
        "operationId": "RemoveParticipant",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{tripId}/optimise": {
      "post": {
        "tags": [
          "Optimisation"
        ],
        "operationId": "Optimise",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimiseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueRunResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{tripId}/runs/{runId}": {
      "get": {
        "tags": [
          "Optimisation"
        ],
        "operationId": "GetRun",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimisationRunDtoWithSolution"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{tripId}/runs/{runId}/pareto": {
      "get": {
        "tags": [
          "Optimisation"
        ],
        "operationId": "GetParetoSolutions",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SolutionDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/trips/{tripId}/participants/{participantId}/calendar.ics": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "operationId": "ParticipantCalendar",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "participantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateNodeDto": {
        "required": [
          "id",
          "participantId",
          "kind",
          "longitude",
          "latitude",
          "walkMins",
          "ptMins",
          "externalId",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "participantId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/CandidateNodeKindDto"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "walkMins": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ptMins": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "externalId": {
            "type": [
              "null",
              "string"
            ]
          },
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CandidateNodeKindDto": {
        "enum": [
          "home",
          "trainStation",
          "busStop",
          "wharf",
          "lightRailStop"
        ]
      },
...
      "DriverRouteDto": {
        "required": [
          "id",
          "driverId",
          "travelMins",
          "orderIndex",
          "stops"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "driverId": {
            "type": "string",
            "format": "uuid"
          },
          "travelMins": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "orderIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StopDto"
            }
          }
        }
      },
      "EnqueueRunResponse": {
        "required": [
          "runId"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ObjectiveWeightsDto": {
        "required": [
          "driveTime",
          "stopCount",
          "walkAndPt",
          "arrivalSpread",
          "fairness"
        ],
        "type": "object",
        "properties": {
          "driveTime": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "stopCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "walkAndPt": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "arrivalSpread": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "fairness": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "OptimisationRunDto": {
        "required": [
          "id",
          "tripId",
          "status",
          "solver",
          "weights",
          "startedAt",
          "completedAt",
          "failureReason",
          "bestSolutionId",
          "stats"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/OptimisationStatus"
          },
          "solver": {
            "$ref": "#/components/schemas/SolverKind"
          },
          "weights": {
            "$ref": "#/components/schemas/ObjectiveWeightsDto"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "failureReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "bestSolutionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "stats": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OptimisationStatsDto"
              }
            ]
          }
        }
      },
      "OptimisationRunDtoWithSolution": {
        "required": [
          "run",
          "solution"
        ],
        "type": "object",
        "properties": {
          "run": {
            "$ref": "#/components/schemas/OptimisationRunDto"
          },
          "solution": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SolutionDto"
              }
            ]
          }
        }
      },
      "OptimisationStatsDto": {
        "required": [
          "wallClock",
          "iterationsOrNodes",
          "bestObjective",
          "lpRelaxation",
          "solver"
        ],
        "type": "object",
        "properties": {
          "wallClock": {
            "pattern": "^-?(\\d+\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string"
          },
          "iterationsOrNodes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "bestObjective": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "lpRelaxation": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "solver": {
            "$ref": "#/components/schemas/SolverKind"
          }
        }
      },
      "OptimisationStatus": {
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ]
      },
      "OptimiseRequest": {
        "required": [
          "weights"
        ],
        "type": "object",
        "properties": {
          "weights": {
            "$ref": "#/components/schemas/ObjectiveWeightsDto"
          },
          "solver": {
            "$ref": "#/components/schemas/SolverKind"
          }
        }
      },
...
      "ParticipantWithNodesDto": {
        "required": [
          "id",
          "tripId",
          "displayName",
          "homeLongitude",
          "homeLatitude",
          "hasCar",
          "seats",
          "preferences",
          "candidateNodes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "homeLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "homeLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "hasCar": {
            "type": "boolean"
          },
          "seats": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesDto"
          },
          "candidateNodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateNodeDto"
            }
          }
        }
      },
      "PreferencesDto": {
        "required": [
          "walkBudgetMins",
          "detourToleranceMins",
          "fairnessWeight"
        ],
        "type": "object",
        "properties": {
          "walkBudgetMins": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detourToleranceMins": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fairnessWeight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "SolutionDto": {
        "required": [
          "id",
          "optimisationRunId",
          "label",
          "objective",
          "objectiveTerms",
          "routes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "optimisationRunId": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "objective": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "objectiveTerms": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          "routes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DriverRouteDto"
            }
          }
        }
      },
      "SolverKind": {
        "enum": [
          "orTools",
          "heuristic"
        ],
        "default": "orTools"
      },
      "StopDto": {
        "required": [
          "id",
          "orderIndex",
          "longitude",
          "latitude",
          "candidateNodeId",
          "estimatedArrival",
          "pickups"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "candidateNodeId": {
            "type": "string",
            "format": "uuid"
          },
          "estimatedArrival": {
            "type": "string",
            "format": "date-time"
          },
          "pickups": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "TripDetailDto": {
        "required": [
          "id",
          "name",
          "destinationName",
          "destinationLongitude",
          "destinationLatitude",
          "departAt",
          "arrivalWindowEarliest",
          "arrivalWindowLatest",
          "ownerId",
          "createdAt",
          "lockedSolutionId",
          "participants"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "destinationName": {
            "type": "string"
          },
          "destinationLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "destinationLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "departAt": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowEarliest": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowLatest": {
            "type": "string",
            "format": "date-time"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lockedSolutionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantWithNodesDto"
            }
          }
        }
      },
...
    }
  },
...
}